@@ -30,10 +30,10 @@
     "migrate:latest": "knex migrate:latest",
     "migrate:latest:prod": "NODE_ENV=production knex migrate:latest",
     "migrate:latest:test": "NODE_ENV=testing knex migrate:latest",
-    "migrate:make": "node ./bin/db.js migrate:make",
-    "migrate:teardown": "node ./bin/db.js migrate:teardown",
-    "migrate:teardown:prod": "NODE_ENV=production node ./bin/db.js migrate:teardown",
-    "migrate:teardown:test": "NODE_ENV=testing node ./bin/db.js migrate:teardown",
+    "migrate:make": "ts-node ./bin/db.ts migrate:make",
+    "migrate:teardown": "ts-node ./bin/db.ts migrate:teardown",
+    "migrate:teardown:prod": "NODE_ENV=production ts-node ./bin/db.ts migrate:teardown",
+    "migrate:teardown:test": "NODE_ENV=testing ts-node ./bin/db.ts migrate:teardown",
     "db:reset": "npm run migrate:teardown && npm run migrate:latest",
     "db:reset:prod": "npm run migrate:teardown:prod && npm run migrate:latest:prod",
     "db:reset:test": "npm run migrate:teardown:test && npm run migrate:latest:test",
@@ -47,51 +47,53 @@
     "start": "node ./build/src/index.js",
     "watch": "tsc -w",
     "build": "tsc && npm run copy",
-    "heroku-postbuild": "npm run build && npm run migrate:latest"
+    "heroku-postbuild": "npm run build && npm run migrate:latest",
+    "exec": "ts-node --files"
   },
   "dependencies": {
     "@types/axios": "^0.14.0",
     "@types/bcrypt": "^3.0.0",
     "@types/cookie": "^0.3.2",
-    "@types/hapi": "^17.6.3",
+    "@types/hapi": "^17.8.2",
     "@types/hapi-auth-jwt2": "^8.0.2",
-    "@types/jest": "^23.3.9",
+    "@types/jest": "^23.3.12",
     "@types/jsonwebtoken": "^8.3.0",
-    "@types/knex": "^0.15.0",
+    "@types/minimist": "^1.2.0",
     "@types/pdfmake": "^0.1.2",
-    "@types/ramda": "^0.25.41",
+    "@types/ramda": "^0.25.46",
     "axios": "^0.18.0",
-    "bcrypt": "^3.0.2",
-    "boom": "^7.2.2",
-    "chalk": "^2.4.1",
+    "bcrypt": "^3.0.3",
+    "boom": "^7.3.0",
+    "chalk": "^2.4.2",
     "cookie": "^0.3.1",
-    "dotenv": "^6.1.0",
-    "hapi": "^17.7.0",
-    "hapi-auth-bearer-token": "^6.1.0",
-    "hapi-auth-jwt2": "^8.1.0",
-    "hapi-pino": "^5.1.0",
+    "dotenv": "^6.2.0",
+    "hapi": "^17.8.1",
+    "hapi-auth-bearer-token": "^6.1.1",
+    "hapi-auth-jwt2": "^8.2.0",
+    "hapi-pino": "^5.2.0",
     "hapi-qs": "^2.0.1",
-    "joi": "^14.0.4",
-    "jsonwebtoken": "^8.3.0",
-    "knex": "^0.15.2",
+    "joi": "^14.3.1",
+    "jsonwebtoken": "^8.4.0",
+    "knex": "^0.16.3",
     "minimist": "^1.2.0",
-    "moment": "^2.22.2",
+    "moment": "^2.23.0",
     "pdfmake": "^0.1.39",
-    "pg": "^7.6.1",
-    "postmark": "^2.2.0",
+    "pg": "^7.7.1",
+    "postmark": "^2.2.4",
     "qrcode": "^1.3.2",
-    "ramda": "^0.25.0",
+    "ramda": "^0.26.1",
     "ts-node": "^7.0.1",
-    "typescript": "^3.1.6"
+    "typescript": "^3.2.2"
   },
   "devDependencies": {
     "@types/mockdate": "^2.0.0",
-    "axios-mock-adapter": "^1.15.0",
-    "factory-girl": "^5.0.3",
+    "axios-mock-adapter": "^1.16.0",
+    "factory-girl": "^5.0.4",
     "jest": "^23.6.0",
+    "minimist": "^1.2.0",
     "mockdate": "^2.0.2",
-    "nodemon": "^1.18.7",
-    "ts-jest": "^23.10.4",
+    "nodemon": "^1.18.9",
+    "ts-jest": "^23.10.5",
     "tslint-config-twine": "github:TwinePlatform/tslint-config-twine#2.1.0"
   }
 }
