@@ -29,7 +29,7 @@ describe('User Model', () => {
     test('get :: no arguments gets all users', async () => {
       const users = await Users.get(trx);
 
-      expect(users.length).toBe(8);
+      expect(users.length).toBe(9);
       expect(users).toEqual(expect.arrayContaining([
         expect.objectContaining({
           name: 'Chell',
@@ -70,6 +70,7 @@ describe('User Model', () => {
           'Barney',
           'Big Boss',
           'Chell',
+          'Companion Cube',
           'Emma Emmerich',
           'GlaDos',
           'Gordon',
@@ -80,7 +81,7 @@ describe('User Model', () => {
 
     test('get :: offset results', async () => {
       const users = await Users.get(trx, { offset: 3, order: ['id', 'asc'] });
-      expect(users.length).toBe(5);
+      expect(users.length).toBe(6);
       expect(users[0].name).toBe('Barney');
     });
 
@@ -194,6 +195,17 @@ describe('User Model', () => {
       } catch (error) {
         expect(error).toBeTruthy();
       }
+    });
+
+    test('add :: can create record with optional age', async () => {
+      const changeset = await factory.build('user');
+      changeset.birthYear = null;
+
+      const user = await Users.add(trx, changeset);
+
+      expect(user).toEqual(expect.objectContaining({
+        ...omit(['password'], changeset),
+      }));
     });
 
     test('destroy :: mark existing record as deleted', async () => {
