@@ -147,7 +147,7 @@ export type User = Readonly<CommonTimestamps & {
   gender: GenderEnum
   disability: DisabilityEnum
   ethnicity: EthnicityEnum
-  birthYear?: Int
+  birthYear?: Int | null
   postCode?: string
   isEmailConfirmed: boolean
   isPhoneNumberConfirmed: boolean
@@ -313,6 +313,8 @@ export type VisitorCollection = UsersBaseCollection & {
     Promise<(Partial<User> & { visits: LinkedVisitEvent[] })[]>
   fromCommunityBusiness: (client: Knex, c: CommunityBusiness, q?: ModelQuery<User>) =>
     Promise<Partial<User>[]>
+  addWithRole: (k: Knex, c: CommunityBusiness, a: Partial<User>) => Promise<User>
+  addAnonymousWithRole: (k: Knex, c: CommunityBusiness, a: Partial<User>) => Promise<User>
 };
 
 export type VolunteerCollection = UsersBaseCollection & {
