@@ -6,7 +6,7 @@ const rndInt = (max: number, min = 0) =>
 export const rndPastDateThisMonth = () =>
   moment().utc()
     .startOf('month')
-    .add(rndInt(moment().date() - 2), 'days')
+    .add(rndInt(moment().date()), 'days')
     .add(rndInt(23), 'hours')
     .add(rndInt(59), 'minutes')
     .add(rndInt(59), 'seconds')
